Fixed broken references and added a quit command
Changed getRoom to return a room reference.
The doors array in room is now initialized to false.
Added SUCCESS_QUIT statuscode, which indicates that the program should be exited. This is passed when the quit command is given.
Removed the noreturn keyword from the run loop, as it can now return.
Updated InputHandler's calling of movePlayer to give enum directions as the parameter.