Worked on room generation.
Began work on a simple maze algorithm, not currently working.
Possibly fixed getRoom, it might not have been broken in the first place.
Changed activeRoom in PlayerState from a reference to a normal object. This fixed some issues.
Added some temporary debug stuff in PlayerAction.cpp and Game.cpp.
Added a generated flag in the Room class, for the maze algorithm to use.
Added the look command for looking around.